+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 ## `transcriptic` Changelog
 
 ## Unreleased
+## v5.5.0
 Added
 - `attachments` attribute on `Dataset`
+
+Fixed
+- Analyzed Dataset content-disposition
 
 ## v5.4.1
 Updated